Finished/Fixed IDS & it's reporting
Turns out I'm an idiot. Fixed some of the problems with IDS tree search
(tie breaking: smaller numbers first, not larger, include the last tile
moved as a possibility (can lead to just moving the same number tile
back and forth)). Fixed the reporting issues as well. The node state is
supposed to be the state of the board, was just printing which tiles
were selected in the order they were selected. Fixed the number of nodes
expanded to be correct, previously wasn't factoring in the starting
state as being a node correctly